backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic-settings==2.8.0
python-dotenv==1.0.1
supabase==2.13.0
httpx==0.27.2
PyJWT==2.10.1
openai==1.66.2
python-multipart==0.0.20
